% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{brenthis}
\alias{brenthis}
\title{Dataset of Butterfly Genotypes}
\format{
vcf file with 13 individuals and 4 markers.
}
\description{
A subset of single nucleotide polymorphisms in butterflies of the genus \emph{Brenthis}.
}
\details{
The data is used to test conversion of genotype data from vcf to diem format with
function \code{vcf2diem}.
}
\examples{
filename <- system.file("extdata", "brenthis.vcf", package = "diemr")
}
