% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.r
\name{testdata}
\alias{testdata}
\title{Dataset of Fish Genotypes}
\format{
vcf file with 92 individuals and 6 markers.
}
\description{
A subset of single nucleotide polymorphisms in fish for testing purposes of
multiallelic markers.
}
\details{
The data is used to test conversion of genotype data from vcf to diem format with
the function \code{vcf2diem}.
}
\examples{
filename <- system.file("extdata", "testdata.vcf", package = "diemr")
}
