% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{readIncludedSites}
\alias{readIncludedSites}
\title{Read genomic sites from a BED-like file}
\usage{
readIncludedSites(includedSites, ChosenSites = "all")
}
\arguments{
\item{includedSites}{Path to a BED-like file, ideally output from \code{vcf2diem}.}

\item{ChosenSites}{Logical or numeric vector specifying sites to retain, or \code{"all"} to keep all sites.}
}
\value{
A data frame containing the selected sites.
}
\description{
Imports a BED-like file and optionally subsets sites.
}
\keyword{internal}
