% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{truncatedLaplace}
\alias{truncatedLaplace}
\title{Truncated Laplace distribution to 95\% of area under the curve, weights scaled to 10/19}
\usage{
truncatedLaplace(x, laplaceScale)
}
\arguments{
\item{x}{A vector of integers, representing centered positions of SNPs}

\item{laplaceScale}{A numeric giving scale of the Laplace distribution.}
}
\description{
Truncated Laplace distribution to 95\% of area under the curve, weights scaled to 10/19
}
\details{
x must be within +-laplaceScale * log(20)
}
