% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{unbiasedWeightedStateChoice}
\alias{unbiasedWeightedStateChoice}
\title{Selects weighted mode of the genomic state}
\usage{
unbiasedWeightedStateChoice(genomicStates, laplaceWeights)
}
\arguments{
\item{genomicStates}{A character vector of genomic states in a given interval}

\item{laplaceWeights}{A numeric vector of weights}
}
\description{
Selects weighted mode of the genomic state
}
